+++ src/bootstrap.jl
@@ -177,7 +177,7 @@ end
 
 """
     parametricbootstrap([rng::AbstractRNG], nsamp::Integer, m::MixedModel{T}, ftype=T;
-        β = coef(m), σ = m.σ, θ = m.θ, hide_progress=false, optsum_overrides=(;))
+        β = coef(m), σ = m.σ, θ = m.θ, progress=true, optsum_overrides=(;))
 
 Perform `nsamp` parametric bootstrap replication fits of `m`, returning a `MixedModelBootstrap`.
 
@@ -193,7 +193,7 @@ performance benefits.
 - `β`, `σ`, and `θ` are the values of `m`'s parameters for simulating the responses.
 - `σ` is only valid for `LinearMixedModel` and `GeneralizedLinearMixedModel` for
 families with a dispersion parameter.
-- `hide_progress` can be used to disable the progress bar. Note that the progress
+- `progress` controls whehter the progress bar is shown. Note that the progress
 bar is automatically disabled for non-interactive (i.e. logging) contexts.
 - `optsum_overrides` is used to override values of [OptSummary](@ref) in the models
 fit during the bootstrapping process. For example, `optsum_overrides=(;ftol_rel=1e-08)`
@@ -210,9 +210,17 @@ function parametricbootstrap(
     σ=morig.σ,
     θ::AbstractVector=morig.θ,
     use_threads::Bool=false,
-    hide_progress::Bool=false,
+    progress::Bool=true,
+    hide_progress::Union{Bool,Nothing}=nothing,
     optsum_overrides=(;),
 ) where {T}
+    if !isnothing(hide_progress)
+        Base.depwarn(
+            "`hide_progress` is deprecated, please use `progress` instead." *
+            "NB: `progress` is a positive action, i.e. `progress=true` means show the progress bar.",
+            :parametricbootstrap; force=true)
+        progress = !hide_progress
+    end
     if σ !== missing
         σ = T(σ)
     end
@@ -232,7 +240,7 @@ function parametricbootstrap(
         "use_threads is deprecated and will be removed in a future release",
         :parametricbootstrap,
     )
-    samp = replicate(n; hide_progress=hide_progress) do
+    samp = replicate(n; progress) do
         simulate!(rng, m; β, σ, θ)
         refit!(m; progress=false)
         # @info "" m.optsum.feval